Add files via upload
I've translated the en.json file to Norwegian. I've tried my best to translate all of the words but some are still remained as engloish cause they sound way better then cause there aren't any norwegian words for them to be honest. I hope this is good enough to be used on the fnbr.co website!